libflux: make flux_msg_fprint() private
Problem: The function flux_msg_fprint() is only used in
one place, internally within libflux.  In addition, the message.c
file is getting long.

Solution: Migrate flux_msg_fprint() and its helper functions to
message_fprint.c.  In addition, rename it to msg_fprint() and make
the function private, since it is not used outside of libflux.

Fixes #3754
